pyproject.toml
[project]
name = "latrace"
version = "0.1.0"
description = "Standalone LATRACE memory service workspace."
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
  "asyncpg>=0.29.0",
  "cryptography>=46.0.3",
  "fastapi>=0.115.0",
  "httpx>=0.27.0",
  "hydra-core>=1.3.2",
  "litellm>=1.45.0",
  "neo4j>=5.28.0",
  "omegaconf>=2.3.0",
  "openai>=1.30.0",
  "pydantic>=2.7.0",
  "PyJWT>=2.10.1",
  "python-dotenv>=1.0.1",
  "PyYAML>=6.0.2",
  "rank-bm25>=0.2.2",
  "requests>=2.31.0",
  "socksio>=1.0.0",
  "uvicorn>=0.30.0",
]

[project.optional-dependencies]
gemini = [
  "google-generativeai>=0.7.2",
]
mcp = [
  "mcp>=1.0.0",
]
milvus = [
  "pymilvus>=2.6.6",
]
local-embeddings = [
  "numpy>=1.26.0",
  "sentence-transformers>=3.0.0",
  "torch>=2.2.0",
  "transformers>=4.41.0",
]
multimodal = [
  "faster-whisper>=1.2.1",
  "hdbscan>=0.8.42",
  "insightface>=0.7.3",
  "numpy>=1.26.0",
  "onnxruntime>=1.24.4",
  "open-clip-torch>=2.24.0",
  "opencv-python>=4.11.0.86",
  "Pillow>=10.4.0",
  "pyannote-audio>=4.0.4",
  "pydub>=0.25.1",
  "python-speech-features>=0.6",
  "torch>=2.2.0",
  "wespeakerruntime>=1.0.1",
]

[dependency-groups]
dev = [
  "bandit>=1.7.9",
  "deptry>=0.23.0",
  "import-linter>=2.0",
  "mypy>=1.10.0",
  "pip-audit>=2.7.3",
  "pytest>=8.2.0",
  "pytest-asyncio>=0.21.0",
  "ruff>=0.6.0",
  "types-PyYAML>=6.0.12.20240808",
  "types-requests>=2.32.0.20240914",
]

[tool.uv]
package = false

[tool.ruff.lint]
ignore = ["E402"]
